yelp-document: Make document-uri property a YelpUri
All implementations of YelpDocument except YelpSimpleDocument keep a
reference to the YelpUri, so move it to the parent class and add a
getter to use the YelpUri from derived classes. This is also useful to
get the YelpUri directly from the document, instead of getting the
document uri string, create a YelpUri and resolve it.